Refactor WeatherParser and WeatherFetcher
This does a little more refactoring of WeatherParser and WeatherFetcher
to isolate everything that is weather provider specific into the
WeatherParser class.  The WeatherFetcher class now deals exclusively
with getting the user's saved settings (which use QSettings).  The
WeatherParser class writes the settings and data to the data structure
used by the weather app (which uses Gio::Settings).

Signed-off-by: Ed Beroset <beroset@ieee.org>