Use open-meteo weather data service
This finishes the change from openweathermap to open-meteo as the
provider of weather data.  This was needed because the existing version
2.5 of the openweathermap API will no longer exist after June 2024.  If
it should ever be necessary to change again, the code for the previous
version is #ifdef'd out and comments describe what is necessary to port
to another weather data provider.

Signed-off-by: Ed Beroset <beroset@ieee.org>